tests: reorganize the base and helper classes for a nicer layout
