+++ tests/util/__init__.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+
+from .base_classes import BaseTestInterface, BaseTestInterfaceProcessing
+from . import mock_filedialog
+from . import mock_messagebox
+from . import mock_warning_window
